gatekeeper/requirements.txt
certifi==2016.9.26
future==0.16.0
python-telegram-bot==5.2.0
requests==2.11.1
RPi.GPIO==0.6.3
urllib3==1.19
pyaudio=0.2.9
pysoundfile=0.8.1
pysoundcard=0.5.2
grovepi
